Внес правки
исправил получения объекта из Optional
тип константы CACHE_SIZE поменял с long на int

+++ src/main/java/faang/school/godbless/BJS2_20336/LruCacheService.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Getter
 public class LruCacheService {
 
-    private static final long CACHE_SIZE = 3;
+    private static final int CACHE_SIZE = 3;
 
     private Map<Integer, Data> dataMap = new HashMap<>();
     private LinkedHashMap<Integer, Data> cache = new LinkedHashMap<>();
